refactor: #3 PR리뷰 적용
