+++ src/main/java/com/JAI/user/controller/UserController.java
@@ -2,8 +2,8 @@ package com.JAI.user.controller;
 
 
 import com.JAI.global.controller.ApiResponse;
-import com.JAI.user.controller.request.UserJoinReq;
-import com.JAI.user.controller.request.UserLoginReq;
+import com.JAI.user.controller.request.UserSignupReq;
+import com.JAI.user.jwt.JWTService;
 import com.JAI.user.service.UserService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -18,12 +18,13 @@ import org.springframework.web.bind.annotation.*;
 public class UserController {
 
     private final UserService userService;
+    private final JWTService jwtService;
 
     // 회원가입(이메일, 비밀번호, 닉네임(중복확인))
-    @PostMapping("/join")
-    public ApiResponse<String> join(@RequestBody @Valid UserJoinReq userJoinReq) {
-        userService.join(userJoinReq);
-        return ApiResponse.ok("Success");//(HttpStatus.CREATED, "success");
+    @PostMapping("/signup")
+    public ApiResponse<String> signup(@RequestBody @Valid UserSignupReq userSignupReq) {
+        userService.signup(userSignupReq);
+        return ApiResponse.ok("Success");
     }
     //회원 정보 수정(프로필 사진, 닉네임(중복확인), 마감시간)
     @PatchMapping("")
@@ -31,24 +32,18 @@ public class UserController {
         return null;
     }
 
-//    @PostMapping("login")
-//    public ApiResponse<String> login(@RequestBody UserLoginReq userLoginReq){
-//        System.out.println("Login Controller");
-//        return null;
-//    }
 
     //회원 정보 조회(마이페이지)
     //프로필 사진, 닉네임, 이메일
-    @GetMapping("/info")
+    @GetMapping()
     public ApiResponse<String> getUserInfo(){
-        System.out.println("GET Mapping");
-        return ApiResponse.ok("Success");
+        return null;
     }
 
     //액세스 토큰 재발급
     @PostMapping("/reissue")
     public ApiResponse<String> reissue(HttpServletRequest request, HttpServletResponse response){
-        userService.reissue(request, response);
+        jwtService.reissue(request, response);
         return ApiResponse.of(HttpStatus.OK,"Reissue Access Token Successfully");
     }
