+++ src/main/java/com/JAI/user/service/UserService.java
@@ -1,12 +1,9 @@
 package com.JAI.user.service;
 
-import com.JAI.global.controller.ApiResponse;
-import com.JAI.user.controller.request.UserJoinReq;
+import com.JAI.user.controller.request.UserSignupReq;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.http.ResponseEntity;
 
 public interface UserService {
-    void join(UserJoinReq userJoinReq);
-    void reissue(HttpServletRequest request, HttpServletResponse response);
+    void signup(UserSignupReq userSignupReq);
 }
